Swap GameOfLife trait to be usize-based
This isn't how you'd typically want to design your trait, but this
way people who are new to Rust don't have to cast to and from usize all
over the place.

The tradeoff is that checking neighbors for liveness gets more
complicated, because you can't subtract -1 from 0 for the first
row/column.